pull from chromium

+++ chrome/browser/history/chrome_history_client.cc
@@ -12,6 +12,7 @@
 #include "chrome/browser/profiles/sql_init_error_message_ids.h"
 #include "chrome/browser/ui/profile_error_dialog.h"
 #include "components/bookmarks/browser/bookmark_model.h"
+#include "components/bookmarks/browser/model_loader.h"
 #include "components/history/core/browser/history_service.h"
 
 ChromeHistoryClient::ChromeHistoryClient(
@@ -78,7 +79,7 @@ void ChromeHistoryClient::BookmarkModelBeingDeleted(
 void ChromeHistoryClient::BookmarkNodeRemoved(
     bookmarks::BookmarkModel* bookmark_model,
     const bookmarks::BookmarkNode* parent,
-    int old_index,
+    size_t old_index,
     const bookmarks::BookmarkNode* node,
     const std::set<GURL>& removed_urls) {
   BaseBookmarkModelObserver::BookmarkNodeRemoved(bookmark_model, parent,
